tests/wpt/tests/trusted-types/eval-function-constructor-untrusted-arguments-and-default-policy-throwing.html

<!DOCTYPE html>
<html>
<head>
  <link rel="help" href="https://w3c.github.io/webappsec-csp/#can-compile-strings">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <meta http-equiv="Content-Security-Policy" content="require-trusted-types-for 'script'">
</head>
<body>
<script>
  let policy = trustedTypes.createPolicy("p", {
    createScript: s => { throw Error('createScript exception'); }
  });
  test(t => {
    assert_throws_js(EvalError, _ => eval("1+2"));
  }, `EvalError thrown if the callback of the default policy throws an error (eval).`);
  test(t => {
    assert_throws_js(EvalError, _ => new Function("return 3;"));
  }, `EvalError thrown if the callback of the default policy throws an error (new Function).`);
</script>
</body>
</html>
